Converted StarRating and ThemeProvider components to TS
