+++ src/components/ThemeProvider.tsx
@@ -1,0 +1,17 @@
+'use client';
+import { ThemeProvider as NextThemesProvider } from 'next-themes';
+import { useEffect, useState, ReactNode } from 'react';
+
+export default function ThemeProvider({ children }: { children: ReactNode }) {
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => setMounted(true), []);
+
+  if (!mounted) return <>{children}</>; // Prevents hydration mismatch
+
+  return (
+    <NextThemesProvider attribute="class" defaultTheme="system">
+      {children}
+    </NextThemesProvider>
+  );
+}
